+++ admin/liste-voitures.php
@@ -27,7 +27,7 @@ $totalPages = ceil($totalCars / _ADMIN_ITEM_PER_PAGE_);
   <section class="connection">
     <div class="connection-header">
       <h1 class="header-titles">Liste Voitures</h1>
-      <a href="/admin/ajouter-voiture.php" class="btn  btn-fill">Ajouter</a>
+      <a href="/admin/ajouter-voiture.php" class="btn btn-fill">Ajouter</a>
     </div>
 
     <div class="connection-wrapper wrapper p-2">
@@ -50,27 +50,25 @@ $totalPages = ceil($totalCars / _ADMIN_ITEM_PER_PAGE_);
           </tr>
           <tbody>
             <?php foreach ($cars as $car) { ?>
-            <tr>
-              <!-- <th scope="row"><?= $car["id"]; ?></th> -->
-              <td><?= $car["code"]; ?></td>
-              <td><?= $car["brand"]; ?></td>
-              <td><?= $car["model"]; ?></td>
-              <td><?= $car["year"]; ?></td>
-              <td><?= $car["price"]; ?></td>
-              <td><?= $car["kilometers"]; ?></td>
-              <td><?= $car["color"]; ?></td>
-              <td><?= $car["gearbox"]; ?></td>
-              <td><?= $car["number_doors"]; ?></td>
-              <td><?= $car["fuel"]; ?></td>
-              <td><?= $car["co"]; ?></td>
-              <td><?= $car["image"]; ?></td>
+              <tr>
+                <!-- <th scope="row"><?= $car["id"]; ?></th> -->
+                <td><?= $car["code"]; ?></td>
+                <td><?= $car["brand"]; ?></td>
+                <td><?= $car["model"]; ?></td>
+                <td><?= $car["year"]; ?></td>
+                <td><?= $car["price"]; ?></td>
+                <td><?= $car["kilometers"]; ?></td>
+                <td><?= $car["color"]; ?></td>
+                <td><?= $car["gearbox"]; ?></td>
+                <td><?= $car["number_doors"]; ?></td>
+                <td><?= $car["fuel"]; ?></td>
+                <td><?= $car["co"]; ?></td>
+                <td><?= $car["image"]; ?></td>
 
-              <td><a href="modifier-voiture.php?id=<?= $car['id'] ?>"><i class="fa-solid fa-pencil"></i></a>
-                | <a href="delete-voiture.php?id=<?= $car['id'] ?>"
-                  onclick="return confirm('Êtes-vous sûr de vouloir supprimer cet employé ?')"><i
-                    class="fa-solid fa-trash-can"></i></a>
-              </td>
-            </tr>
+                <td><a href="modifier-voiture.php?id=<?= $car['id'] ?>"><i class="fa-solid fa-pencil"></i></a>
+                  | <a href="delete-voiture.php?id=<?= $car['id'] ?>" onclick="return confirm('Êtes-vous sûr de vouloir supprimer cet employé ?')"><i class="fa-solid fa-trash-can"></i></a>
+                </td>
+              </tr>
             <?php } ?>
           </tbody>
         </table>
@@ -78,13 +76,15 @@ $totalPages = ceil($totalCars / _ADMIN_ITEM_PER_PAGE_);
       <nav aria-label="Page navigation voitures">
         <ul class="pagination">
           <?php if ($totalPages > 1) { ?>
-          <?php for ($i = 1; $i <= $totalPages; $i++) { ?>
-          <li class="page-item">
-            <a class="page-link <?php if ($i == $page) { echo " active";} ?>" href="?page=<?php echo $i; ?>">
-              <?php echo $i; ?>
-            </a>
-          </li>
-          <?php } ?>
+            <?php for ($i = 1; $i <= $totalPages; $i++) { ?>
+              <li class="page-item">
+                <a class="page-link <?php if ($i == $page) {
+                                      echo " active";
+                                    } ?>" href="?page=<?php echo $i; ?>">
+                  <?php echo $i; ?>
+                </a>
+              </li>
+            <?php } ?>
           <?php } ?>
         </ul>
       </nav>
@@ -93,5 +93,5 @@ $totalPages = ceil($totalCars / _ADMIN_ITEM_PER_PAGE_);
   <!-- END connection  -->
 </div>
 <?php
-require_once __DIR__."/templates/footer-admin.php";
+require_once __DIR__ . "/templates/footer-admin.php";
 ?>
